requirements.txt
Authlib
blinker
cachelib
certifi
cffi
charset-normalizer
click
cryptography
dnspython
Flask
Flask-Login
Flask-Mail
Flask-OAuthlib
Flask-PyMongo
Flask-Session
idna
itsdangerous
Jinja2
MarkupSafe
msgspec
oauthlib
pycparser
pymongo
python-dotenv
requests
requests-oauthlib
urllib3
Werkzeug
python-dateutil
python-barcode
pillow
transformers
reportlab
folium
gunicorn
pytest
pytest-cov
pytest-flask
selenium

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-selenium==4.0.1
selenium==4.15.2
flask-testing==0.8.1
